backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langgraph==0.0.20
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pytesseract==0.3.10
easyocr==1.7.0
# PyMuPDF==1.23.14  # Windows에서 컴파일 문제로 제외
PyPDF2==3.0.1
pdfplumber==0.10.3
llama-parse==0.4.0
openai==1.3.7
anthropic==0.7.8
sentence-transformers==2.2.2
faiss-cpu==1.12.0
chromadb==0.4.18
numpy==1.24.3
pandas==2.1.4
aiofiles==23.2.1
httpx==0.25.2
pydantic-settings==2.1.0
Pillow==10.1.0
opencv-python==4.8.1.78
langchain-experimental==0.0.47
